requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
notion-client==2.2.1
python-multipart==0.0.6
python-dotenv==1.0.0
jinja2==3.1.2
reportlab==4.0.7
requests==2.31.0
pydantic==2.5.0
aiofiles==23.2.0

# For local AI (optional - can be added later)
# ollama-python==0.1.7

# For PDF generation alternatives
weasyprint==60.2

# For testing
pytest==7.4.3
httpx==0.25.2
